# services/forumline-api/openapi.yaml
openapi: 3.0.3
info:
  title: Forumline API
  description: The Forumline platform hub — DMs, forums, memberships, notifications, calls, and presence.
  version: 1.0.0
servers:
  - url: https://app.forumline.net
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceKey:
      type: http
      scheme: bearer
      description: Service account PAT for forum-to-platform calls
  schemas:
...
    ProfileSearchResult:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
...
paths:
...
  /api/profiles/search:
    get:
      operationId: searchProfiles
      summary: Search user profiles by username or display name
      tags: [Identity]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileSearchResult'
...